+++ wpyprint/__init__.py
@@ -1,2 +1,3 @@
 from .PPContext import PPContext
+from .print import print
 from .utils import *